requirements.txt
Django==1.9.10
Django-Select2==4.3.2
Pillow==3.3.1
Unidecode==0.04.19
YURL==0.13
aldryn-search==0.2.12
argparse==1.2.1
awscli
backport-collections==0.1
bcrypt==3.1.1
biopython==1.68
bleach==1.4.3
cmsplugin-filer==1.1.3
django-activity-stream==0.6.3
django-appconf==1.0.2
django-appdata==0.1.5
django-bootstrap3==7.1.0
django-classy-tags==0.8.0
django-cms==3.3.2
django-contrib-comments==1.7.3
django-compressor==1.5
django-el-pagination==3.0.1
django-embed-video==1.1.0
django-formtools==1.0
django-filer==1.2.5
django-haystack==2.5.0
django-libsass==0.7
django-localflavor==1.3
django-parler==1.6.5
django-polymorphic==0.8.1
django-redis==4.5.0
django-reversion==1.10.2
django-sekizai==0.10.0
django-taggit==0.21.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.4
djangocms-blog==0.8.8
djangocms-column==1.7.0
djangocms-link==2.0.1
djangocms-text-ckeditor==3.2.1
easy-thumbnails==2.3
elasticsearch==2.3.0
html5lib==0.9999999
ipython
iso8601==0.1.11
libsass==0.11.1
orcid==0.7.0
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.6.1
pyzotero==1.1.19
requests==2.11.1
tzlocal==1.2.2
